Move cacheSnapshot and putCacheEntry test helpers to export_test.go
